requirements.txt
streamlit
pillow
transformers
torch